메서드들끼리 변환의 용의성 + 메모리 사용 줄이기 위해서 Long을 int형으로 변경

+++ AlgoShuffer/src/main/java/com/example/algoshuffer/entity/User.java
@@ -25,10 +25,10 @@ public class User {
     private String username;
 
     @Column(name = "rating")
-    private Long rating;
+    private int rating;
 
     @Column(name = "rank")
-    private Long rank;
+    private int rank;
 
     @OneToMany(mappedBy = "user")
     @JsonIgnore
